manage conversation and message list updates whenever the app is opened

+++ Pulse/Utils/TimeUtils.swift
@@ -29,8 +29,8 @@ extension Int64 {
 }
 
 extension Date {
-    var millisecondsSince1970:Int {
-        return Int((self.timeIntervalSince1970 * 1000.0).rounded())
+    var millisecondsSince1970:Int64 {
+        return Int64((self.timeIntervalSince1970 * 1000.0).rounded())
     }
     
     init(milliseconds:Int64) {
